Further network coding efficiency testing
I think I understand enough now to actually choose the right parameters and use it appropriately! Efficiency is now ~120% (for both number of packets and data size (including coefficients). Yet to be implemented... it seems as though I should extend the symbol size to the packet size for best results, which makes sense from a packet-wise view of the network as well.